# optimizing the spawn placement algorithm