So I thought the new spawn placement algorithm is pretty cool, but it's slow. It's only done once per game and premature optimization is the mother of all evils, but this is fairly easy.
Instead of calculating the spawn ratings for each tile for every attempt to place the civs, just calculate it once and store it in a HashMap.

**Does it help?**
It does, especially on large maps. Times are on my PC.
On standard settings, (Medium map, 4 civs, 6 CS):
Time to generate starting places went from 275 to 213 ms

Huge, 9 civs, 18 CS:
Went from 3314 to 1055 ms

Custom 60 radius, 9 civs, 28 CS:
Went from 10056 to 3193 ms