@ravignir 2021-07-24
Anyway, I am 90% sure in civ5, the civ placing algorithm worked before placing resource.