@SimonCeder 2021-07-24
It's possible. It does seem like all civs get at least 2 iron in civ V. I'll take a closer look at some point.