@yairm210 2021-07-24
This looks like a good fix to the current algorithm, even if the algorithm itself may need tweaking.